extract color utils
