+++ src/assets/theme/helpers/index.ts
@@ -13,4 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from 'src/assets/theme/helpers/border-utils';
+export * from './border-utils';
+export * from './color-utils';
